gps: Combine source mgr test files
